+++ upkie/__init__.py
@@ -6,10 +6,12 @@
 """Python module to control Upkie wheeled biped robots."""
 
 from . import envs, model, utils
+from .model import Model
 
 __version__ = "11.0.0"
 
 __all__ = [
+    "Model",
     "envs",
     "model",
     "utils",